requirements.txt
sentence-transformers>=3.0.0
scikit-learn>=1.3.0
numpy>=1.25.0
pandas>=2.1.0
scipy>=1.11.0
streamlit>=1.34.0
spacy>=3.7.0
python-docx>=1.0.1
pymupdf>=1.23.0
pdfminer.six>=20231228
